+++ packages/element/src/bounds.ts
@@ -65,7 +65,6 @@ import type {
   ExcalidrawLinearElement,
   ExcalidrawRectanguloidElement,
   ExcalidrawTextElementWithContainer,
-  NonDeleted,
 } from "./types";
 
 export type RectangleBox = {
@@ -1081,33 +1080,6 @@ export const getClosestElementBounds = (
   });
 
   return getElementBounds(closestElement, elementsMap);
-};
-
-export interface BoundingBox {
-  minX: number;
-  minY: number;
-  maxX: number;
-  maxY: number;
-  midX: number;
-  midY: number;
-  width: number;
-  height: number;
-}
-
-export const getCommonBoundingBox = (
-  elements: ExcalidrawElement[] | readonly NonDeleted<ExcalidrawElement>[],
-): BoundingBox => {
-  const [minX, minY, maxX, maxY] = getCommonBounds(elements);
-  return {
-    minX,
-    minY,
-    maxX,
-    maxY,
-    width: maxX - minX,
-    height: maxY - minY,
-    midX: (minX + maxX) / 2,
-    midY: (minY + maxY) / 2,
-  };
 };
 
 /**
